+++ gradle.properties
@@ -1,0 +1,1 @@
+kotlin.code.style=official
