+++ src/main/kotlin/io/provenance/client/gas/estimators/CosmosSimulation.kt
@@ -11,16 +11,16 @@ import kotlin.math.ceil
  * Cosmos simulation gas estimation. Must be used when interacting with pbc 1.7 or lower.
  * TODO - Remove once mainnet.version > 1.8
  */
-internal fun cosmosSimulationGasEstimator(gasPrices: GasPrices) : PbGasEstimator = {
-	{ tx, adjustment ->
-		val price = gasPrices()
-		val sim = cosmosService.simulate(
-			ServiceOuterClass.SimulateRequest.newBuilder()
-				.setTxBytes(tx.toByteString())
-				.build()
-		)
-		val limit = ceil(sim.gasInfo.gasUsed * adjustment).toLong()
-		val feeAmount = ceil(limit * price.amount.toDouble()).toLong()
-		GasEstimate(limit, listOf(feeAmount.toCoin(price.denom)))
-	}
+internal fun cosmosSimulationGasEstimator(gasPrices: GasPrices): PbGasEstimator = {
+    { tx, adjustment ->
+        val price = gasPrices()
+        val sim = cosmosService.simulate(
+            ServiceOuterClass.SimulateRequest.newBuilder()
+                .setTxBytes(tx.toByteString())
+                .build()
+        )
+        val limit = ceil(sim.gasInfo.gasUsed * adjustment).toLong()
+        val feeAmount = ceil(limit * price.amount.toDouble()).toLong()
+        GasEstimate(limit, listOf(feeAmount.toCoin(price.denom)))
+    }
 }
