+++ src/main/kotlin/io/provenance/client/grpc/PbClient.kt
@@ -5,7 +5,6 @@ import cosmos.tx.v1beta1.ServiceOuterClass
 import cosmos.tx.v1beta1.TxOuterClass
 import cosmos.tx.v1beta1.TxOuterClass.TxBody
 import io.grpc.netty.shaded.io.grpc.netty.NettyChannelBuilder
-import io.provenance.client.TestnetFeaturePreview
 import io.provenance.client.protobuf.extensions.getBaseAccount
 import io.provenance.msgfees.v1.QueryParamsRequest
 import java.io.Closeable
@@ -70,7 +69,7 @@ open class PbClient(
     val metadataClient = io.provenance.metadata.v1.QueryGrpc.newBlockingStub(channel)
     val mintClient = cosmos.mint.v1beta1.QueryGrpc.newBlockingStub(channel)
 
-    @TestnetFeaturePreview
+    // @TestnetFeaturePreview
     val msgFeeClient = io.provenance.msgfees.v1.QueryGrpc.newBlockingStub(channel)
 
     val nameClient = io.provenance.name.v1.QueryGrpc.newBlockingStub(channel)
@@ -81,10 +80,10 @@ open class PbClient(
     val upgradeClient = cosmos.upgrade.v1beta1.QueryGrpc.newBlockingStub(channel)
     val wasmClient = cosmwasm.wasm.v1.QueryGrpc.newBlockingStub(channel)
 
-    @TestnetFeaturePreview
+    // @TestnetFeaturePreview
     val nodeFeeParams = lazy { msgFeeClient.params(QueryParamsRequest.getDefaultInstance()).params }
 
-    @TestnetFeaturePreview
+    // @TestnetFeaturePreview
     val nodeGasPrice = lazy { nodeFeeParams.value.floorGasPrice.amount.toDouble() }
 
     fun baseRequest(
